Fix MemberPackage: 패키지 구조 변경 및 안쓰는 import 제거
