+++ README.md
@@ -84,7 +84,6 @@ code with me 를 활용하여 어렵거나 힘든 기능을
 - 삭제
 
 ## 프로젝트 기능 요구 사항
-
 "*" : 변경이 필요한 요구사항입니다.
 
 ### ✏️회원
