+++ src/main/java/study/tdd/simpleboard/api/member/entity/MemberRepository.java
@@ -1,4 +1,4 @@
-package study.tdd.simpleboard.api.member.signup;
+package study.tdd.simpleboard.api.member.entity;
 
 import study.tdd.simpleboard.api.member.entity.Member;
